auth-service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
PyJWT
Werkzeug
PyJWT                  2.10.1